requirements.txt
langchain
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf~=3.19.0
langchain_experimental
mysql-connector-python
pymysql
sentence-transformers
chromadb==0.4.15
ipykernel
google-generativeai